Move components related to action form to separate dir
