+++ src/components/campaign/action/ResponseWidget.jsx
