+++ src/components/campaign/action/ResponseWidget.scss
